pyproject.toml
[project]
name = "paper-check"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flet[all]>=0.28.3",
    "flet-cli>=0.28.3",
    "python-docx>=1.2.0",
]
